// src/app/shared/models/gateway.ts
export class Gateway {
  id_gateway: number;
  id_contract: number;
  id_gateway_type: number;
  id_owner: number;
  id_geograph: number;
  id_node: number;

  code_contract: string;
  code_owner: string;
  code_gateway_type: string;
  code_geograph: string;

  n_coordinate: number;
  e_coordinate: number;
  price: number;
  comments: string;

  dateedit: string;
  useredit: number;
}